requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
astroid==2.1.0
autopep8==1.4.3
certifi==2018.11.29
chardet==3.0.4
coverage==4.5.2
coveralls==1.5.1
dj-database-url==0.5.0
django-cors-middleware==1.3.1
django-extensions==2.1.4
django-nose==1.4.6
django==2.1.5
djangorestframework==3.9.1
docopt==0.6.2
gunicorn==19.9.0
idna==2.8
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
nose==1.3.7
pillow==5.4.1
psycopg2-binary==2.7.6.1
psycopg2==2.7.6.1
pycodestyle==2.4.0
pyjwt==1.7.1
pylint==2.2.2
python-decouple==3.1
pytz==2018.9
requests==2.21.0
six==1.12.0
typed-ast==1.2.0
urllib3==1.24.1
whitenoise==4.1.2
wrapt==1.11.0
